src/ingest/requirements.txt
google-cloud-storage>=2.15.0
google-cloud-bigquery>=3.18.0
google-cloud-logging>=3.10.0
python-dateutil>=2.8.2
functions-framework>=3.0.0
google-cloud-bigquery>=3.18.0
pytest>=8.2.0
pytest-mock>=3.14.0
black>=23.12
ruff>=0.3